Added automatic assertion type detection
As of this commit, formatting codes no longer have to be specified for
comparison-type assertions.  This is accomplished through a C11 _Generic
selector that matches the type of the asserted value to a specialised
pretty-print function that writes to the failure message buffer.  As
such, this library is now requires C11 compiler support, but given that
it was really only compatible with GCC and Clang anyway, that shouldn't
be much of an issue.  Furthermore, compile times are now slightly longer
due to the use of generic metaprogramming, but not so much so as to be
an impediment to the usefulness of this library.

I also documented the compiler requirements in the README and updated
the example test suite to reflect the loss of formatting codes for
assertions.