Rewrite: Integrate MSVC-compatible test.h from Prota
This commit introduces a version of test.h that I rewrote from
scratch for a project called Prota, that is now almost definitely
abandoned.  This version has a few key differences from older
versions of the library, the largest of which is that it provides
MSVC support (tested back through MSVC 6) at the expense of no longer
performing signal handling since there is no cross-platform way to
do it.  The other files in this repo will need to be migrated to the
new test syntax, which will be done shortly.  Other than that, I
consider this version of test.h to be a strict upgrade over the old
gcc/clang-only version and am planning to make it the primary offering
going forward.